% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ife.R
\name{print.ife}
\alias{print.ife}
\title{Print an ife object}
\usage{
\method{print}{ife}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{"ife"}}

\item{digits}{number of significant digits (default 4)}

\item{...}{unused}
}
\value{
\code{x} invisibly.
}
\description{
Print an ife object
}
